refactor(emotes-service): move Pulumi Go code into infra directory
Separates infrastructure code from future application code.
Uses Pulumi.yaml `main: infra` field so scripts stay unchanged.
