+++ apps/emotes-service/infra/util/git.go
