+++ apps/emotes-service/infra/util/git_test.go
