+++ apps/emotes-service/infra/util/stack.go
